ExternalSortingTable custom class
