+++ jdi-light-vuetify-tests/src/test/java/io/github/epam/vuetify/tests/complex/DataTablesTests.java
@@ -1,24 +1,15 @@
 package io.github.epam.vuetify.tests.complex;
 
-import static com.epam.jdi.light.asserts.core.SoftAssert.jdiAssert;
-import static com.jdiai.tools.Timer.waitCondition;
-import static io.github.com.enums.TableTestData.DONUT;
-import static io.github.com.enums.TableTestData.ECLAIR;
-import static io.github.com.enums.TableTestData.ECLAIR_CALORIES;
-import static io.github.com.enums.TableTestData.FROZEN_YOGURT;
-import static io.github.com.enums.TableTestData.HONEYCOMB;
-import static io.github.com.enums.TableTestData.HONEYCOMB_CALORIES;
-import static io.github.com.enums.TableTestData.ICE_CREAM_SANDWICH;
-import static io.github.com.enums.TableTestData.JELLY_BEAN;
-import static io.github.com.enums.TableTestData.KITKAT;
-import static io.github.com.enums.TableTestData.KITKAT_CALORIES;
-import static io.github.com.StaticSite.dataTablesPage;
-import static io.github.com.pages.DataTablesPage.*;
-
 import io.github.epam.TestsInit;
 import org.hamcrest.Matchers;
 import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Test;
+
+import static com.epam.jdi.light.asserts.core.SoftAssert.jdiAssert;
+import static com.jdiai.tools.Timer.waitCondition;
+import static io.github.com.StaticSite.dataTablesPage;
+import static io.github.com.enums.TableTestData.*;
+import static io.github.com.pages.DataTablesPage.*;
 
 public class DataTablesTests extends TestsInit {
     // TODO: Move table description to jdi-light-vuetify-tests project and implement DataTable as general class
